ref(lib): Moves signatures and invoices to their own modules
Each of these types and their implementations were getting pretty big
so it made sense to move them to their own modules and reexport them
at the top level
